+++ client/src/components/Other/GroupComponent.js
@@ -1,6 +1,4 @@
 import React, { useState } from "react";
-import {deleteCustomAsset} from "../../utils/portfolio";
-import {useAuth0} from "@auth0/auth0-react";
 
 const GroupComponent = ({
   account,
@@ -10,19 +8,11 @@ const GroupComponent = ({
   getTransactionsFunc,
   baseSymbol,
   custom_asset,
+  deleteCustomAssetFunc,
 }) => {
   let amount = 0;
   let currency = "";
   let base_currency = 0;
-
-  const {getAccessTokenSilently} = useAuth0();
-
-  const deleteCustomAssetFunc = async (asset, asset_type) => {
-    const token = await getAccessTokenSilently();
-    await deleteCustomAsset(token, asset, asset_type)
-    window.sessionStorage.clear();
-    window.location.reload()
-  }
 
   if (account.data.balanceData) {
     amount = Number(account.data.balanceData.amount);
